backend/requirements.txt
uvicorn==0.23.2
fastapi==0.109.1
alembic==1.12.0
psycopg2==2.9.9
asyncpg==0.28.0
sqlalchemy[asyncio]==2.0.20
PyPDF2==3.0.1
pgvector==0.3.5
sentence-transformers==3.2.0
python-multipart==0.0.12
email_validator==2.2.0
bcrypt==4.2.1
PyJWT==2.10.0
litellm==1.51.0
